+++ packages/vue/src/examples/pages/checkout/_internal/ConfirmOrder.vue
