+++ packages/vue/src/examples/pages/checkout/_internal/index.js
@@ -1,14 +1,14 @@
 import PersonalDetails from "./PersonalDetails.vue";
 import Shipping from "./Shipping.vue";
 import Payment from "./Payment.vue";
-import ReviewOrder from "./ReviewOrder.vue";
+import ConfirmOrder from "./ConfirmOrder.vue";
 import OrderSummary from "./OrderSummary.vue";
 import OrderReview from "./OrderReview.vue";
 export {
   PersonalDetails,
   Shipping,
   Payment,
-  ReviewOrder,
+  ConfirmOrder,
   OrderSummary,
   OrderReview
 };
